# Video optimization (Linux CPU 100% usage, Issue #51)