- render_line:
  - Recalculating the XRGB color from palette per line only when changed, not when writing to the framebuffer.
  - Split up the loop over pixels in a line into discrete stages:
      1) Calculate color indices.
      2) Add border.
      3) Look up XRGB color & write to framebuffer.
      4) Apply NTSC overscan.
   - Calculate 8 color indices per iteration (was only maybe a 50ms improvement for quite a bit more complexity... probably be much nicer if flatted by hand to SIMD intrinsics)

- render_layer_line:
  - Calculate VRAM range to read in, and read in the range rather than several individual bytes when the range lies within VRAM, otherwise fallback to individual byte reads.

Timing video_step over 300 frames (158025 calls):
Baseline:
    2477.246284 ms total, 0.015676 ms avg,
    **8.257487613 ms/frame**

Plus optimized render_line:
    1848.739386 ms total, 0.011699 ms avg,
    **6.16246462 ms/frame**

Plus optimized render_layer_line:
    1673.707724 ms total, 0.010591 ms avg,
    **5.579025747 ms/frame**

Additional testing with @indigodarkwolf's x16-racer, using "-log S" specifically:
```
Before (Settles at 88-90%):
Load: 100%
Rendering is behind -58 frames.
Load: 100%
Rendering is behind -58 frames.
Load: 100%
Rendering is behind -58 frames.
Load: 100%
Rendering is behind -58 frames.
Load: 100%
Rendering is behind -59 frames.
Load: 100%
Rendering is behind -59 frames.
Load: 100%
Rendering is behind -59 frames.
Load: 100%
Rendering is behind -59 frames.
Load: 100%
Rendering is behind -59 frames.
```
After:
```
Load: 100%
Load: 88%
Load: 94%
Load: 94%
Load: 82%
Load: 100%
Load: 88%
Load: 88%
Load: 88%
Load: 100%
Load: 82%
Load: 100%
Load: 88%
Load: 94%
Load: 82%
```

Test CPU is an Intel(R) Core(TM) i5-5250U CPU @ 1.60GHz.