@neilogd 2019-10-17
I'd be tempted to push a bit more and squeeze some more juice, but at this stage in the project I'd rather wait until the hardware is unlikely to change much. Don't want to do optimization work that has a high chance of being obsolete in a revision or two if can be avoided :)