@mist64 2019-11-02
I'm planning to release r34 *without* this change later today (for stability), and have a look at this asap... :/